backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy>=2.0.0
scipy>=1.11.4
librosa>=0.10.0
openai-whisper>=20231117
torch>=2.1.1
torchaudio>=2.1.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
